Fix type mismatch: Change Int to Long in ScannerViewModel.kt

+++ app/src/main/java/com/example/struku/presentation/scan/ScannerViewModel.kt
@@ -28,7 +28,7 @@ class ScannerViewModel @Inject constructor(
                 if (receiptId != null) {
                     _state.update { it.copy(
                         isLoading = false,
-                        scannedReceiptId = receiptId.toInt()
+                        scannedReceiptId = receiptId
                     ) }
                 } else {
                     _state.update { it.copy(
@@ -53,5 +53,5 @@ class ScannerViewModel @Inject constructor(
 data class ScannerState(
     val isLoading: Boolean = false,
     val error: String? = null,
-    val scannedReceiptId: Int? = null
+    val scannedReceiptId: Long? = null
 )
